test: 테스트 코드 리팩토링
- depreciated 메서드 변경
- 코드 스타일 정리
- 로그 테스트 추가
